+++ src/domain/devices/synth_device.hpp
@@ -84,8 +84,8 @@ public:
     void setVco1Waveform(PolyBLEPOscillator::Waveform wave);
     int vco1Octave() const;
     void setVco1Octave(int octave);
-    float vco1Pitch() const;
-    void setVco1Pitch(float pitch);
+    int vco1Pitch() const;
+    void setVco1Pitch(int pitch);
     float vco1Shape() const;
     void setVco1Shape(float shape);
     bool vco1Sync() const;
@@ -96,8 +96,8 @@ public:
     void setVco2Waveform(PolyBLEPOscillator::Waveform wave);
     int vco2Octave() const;
     void setVco2Octave(int octave);
-    float vco2Pitch() const;
-    void setVco2Pitch(float pitch);
+    int vco2Pitch() const;
+    void setVco2Pitch(int pitch);
     float vco2Shape() const;
     void setVco2Shape(float shape);
     bool vco2Sync() const;
@@ -221,13 +221,13 @@ private:
     // Internal parameter storage
     PolyBLEPOscillator::Waveform m_vco1Waveform { PolyBLEPOscillator::Waveform::Saw };
     int m_vco1Octave { 0 };
-    float m_vco1Pitch { 0.0f };
+    int m_vco1Pitch { 0 };
     float m_vco1Shape { 0.0f };
     bool m_vco1Sync { false };
 
     PolyBLEPOscillator::Waveform m_vco2Waveform { PolyBLEPOscillator::Waveform::Saw };
     int m_vco2Octave { 0 };
-    float m_vco2Pitch { 0.0f };
+    int m_vco2Pitch { 0 };
     float m_vco2Shape { 0.0f };
     bool m_vco2Sync { false };
